+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.2.0 (Next)
 
+* [#5](https://github.com/dblock/ruby-link-checker/pull/5): Fixed false positive redirect loop - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.0 (2023-04-02)